peterju4_project/requirements.txt
Flask==2.3.2
google-cloud-datastore==2.8.2
werkzeug==2.3.3
python-jose
flask-cors
six
python-dotenv
requests
authlib
protobuf==3.20.*